+++ wp-etransaction/admin/products-post.php
@@ -11,9 +11,10 @@ add_action('admin_post_add_product', function () {
     $price = $_REQUEST['price'];
     $active = isset($_REQUEST['active']);
     $free_amount = isset($_REQUEST['free_amount']);
+    $category = $_REQUEST['category'];
 
     $productDb = ProductDB::get_instance();
-    $productDb->insert($name, $price, $active, $free_amount);
+    $productDb->insert($name, $price, $active, $free_amount, $category);
 
     redirect();
 });
@@ -46,9 +47,10 @@ add_action('admin_post_edit_product', function () {
     $price = $_REQUEST['price'];
     $active = isset($_REQUEST['active']);
     $free_amount = isset($_REQUEST['free_amount']);
+    $category = $_REQUEST['category'];
 
     $productDb = ProductDB::get_instance();
-    $result = $productDb->update($product_id, $name, $price, $active, $free_amount);
+    $result = $productDb->update($product_id, $name, $price, $active, $free_amount, $category);
     if ($result === false) {
         echo "Pas marche: " . $productDb->pkoi();
         die();