+++ ChangeLog.md
@@ -6,6 +6,7 @@
     * Add free amount for product. e.g. for donation.
     * Remove hard coded string for translation key
     * Change translation functions call due to a misunderstood how it works.
+    * Add product cateogry
     
 * v1.0.0: 
     * Initial Release